requirements.txt
jtextfsm
jinja2
netaddr
pyYAML
napalm-yang
